Localize thermometer address reference
There is a bug in the firmware where occasionally we get the same
temperature for multiple devices on the wire. My suspicion is this is
related to a stale thermometer address being sent as the wrong
temperature index. No idea if this is going to be the right fix, but
worth trying out.